pyproject.toml
[tool.poetry]
name = "j63"
version = "0.1.0"
description = ""
authors = ["Jennifer Wilcox <jennifer@nitori.org>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
black = "^25.1.0"
ruff = "^0.9.9"
vsg = "^3.31.0"
tomli-w = "^1.2.0"
pyserial = "^3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tools.ruff.lint]
extend-select = ["E", "F", "I", "N", "PL"]
